Added solution to probelm 34 (note that it is VERY slow).

+++ euler_util.py
@@ -16,4 +16,4 @@ def is_prime(n):
     return True
 
 def digits(n):
-    return [long(x) for x in str(n)]
+    return [int(x) for x in str(n)]
